+++ R/transition.R
@@ -141,7 +141,7 @@ trans_apply <- function (idf, ver, keep_all) {
                 " From  Ver: ", vers[i], "\n",
                 "Toward Ver: ", vers[i + 1L]
             )
-            idf <- with_speed(trans_funs[[funs[i]]](idf))
+            idf <- without_checking(trans_funs[[funs[i]]](idf))
             verbose_info("[", vers[i], " --> ", vers[i + 1L], "] SUCCEEDED.\n")
         }
         idf
@@ -155,7 +155,7 @@ trans_apply <- function (idf, ver, keep_all) {
                 " From  Ver: ", vers[i], "\n",
                 "Toward Ver: ", vers[i + 1L]
             )
-            res[[i + 1L]] <- with_speed(trans_funs[[funs[[i]]]](res[[i]]))
+            res[[i + 1L]] <- without_checking(trans_funs[[funs[[i]]]](res[[i]]))
             verbose_info("[", vers[i], " --> ", vers[i + 1L], "] SUCCEEDED.\n")
         }
         nms <- paste0(stri_sub(funs, 6L, 6L), ".", stri_sub(funs, 7L, 7L))